+++ src/main/scala/model/SCUser.scala
@@ -22,9 +22,9 @@ case class SCUser(userId: String,
   def scrollRate: BigDecimal = BigDecimal(scrollCount*100.00/totalKills).setScale(2, RoundingMode.HALF_UP)
   def donutRate: BigDecimal = BigDecimal(donutCount*100.00/totalKills).setScale(2, RoundingMode.HALF_UP)
 
-  def expectedSuncrystals: Double = totalKills/6
-  def expectedScrolls: Double = (totalKills*247)/300
-  def expectedDonuts: Double = totalKills/100
+  def expectedSuncrystals: Long = totalKills/6
+  def expectedScrolls: Long = (totalKills*247)/300
+  def expectedDonuts: Long = totalKills/100
 
   def suncrystalOffset: Double = ((scCount-expectedSuncrystals)*100.00)/expectedSuncrystals
   def scrollsOffset: Double =  ((scCount-expectedScrolls)*100.00)/expectedScrolls
